move UnitFeaturesPage
