+++ frontend/src/employee-frontend/App.tsx
@@ -28,7 +28,6 @@ import LoginPage from './components/LoginPage'
 import MobilePairingModal from './components/MobilePairingModal'
 import PersonalMobileDevicesPage from './components/PersonalMobileDevicesPage'
 import Reports from './components/Reports'
-import UnitFeaturesPage from './components/UnitFeaturesPage'
 import UnitPage from './components/UnitPage'
 import Units from './components/Units'
 import WelcomePage from './components/WelcomePage'
@@ -123,6 +122,7 @@ import DiscussionReservationSurveyWrapper from './components/unit/tab-calendar/d
 import DiscussionReservationSurveysPage from './components/unit/tab-calendar/discussion-surveys/DiscussionSurveysPage'
 import CreateUnitPage from './components/unit/unit-details/CreateUnitPage'
 import UnitDetailsPage from './components/unit/unit-details/UnitDetailsPage'
+import UnitFeaturesPage from './components/unit-features/UnitFeaturesPage'
 import VoucherValueDecisionPage from './components/voucher-value-decision/VoucherValueDecisionPage'
 import VoucherValueDecisionsPage from './components/voucher-value-decisions/VoucherValueDecisionsPage'
 import { queryClient, QueryClientProvider } from './query'
